Fix "everything passes, travis fails"
Travis support pointed to buildApps as a possible culprit. After
investigating, I found out `buildSandcastle` is returning a stream that
is passed to Promise.join. That means the promise resolves before the
stream is finished. Instead it should be returning a promise created from
the stream.

This is a good change whether it fixes the issue or not, since the code
in master is clearly wrong.